chore: avoid using index files in extension
